fix: types and addTask call
