+++ types/service.ts
@@ -1,0 +1,10 @@
+export interface AddTaskInput {
+  text: string;
+  tagIds?: string[];
+  dueDate?: string;
+  priority?: number;
+}
+
+export type ServiceAddTaskInput = {
+  userId: string;
+} & AddTaskInput;
